Create TurtleBakeTextureAssigner
This script is written in Maya Embedded Language. It was created by Ivan Spalla and published on 9.8.2015. It is free for use under Creative Commons Licensing. It has been tested in Maya 2014. 

Use this script once you have baked items in a scene using the TURTLE renderer within Maya. Please use the shape flag in the name of your bakes, as this script is heavily dependant on it. Select objects 
that need to have textures applied, specificy the file folder in which your baked textures can be found, specificy the file prefix, and simply click Assign.

Any questions can be directed to spallam42@gmail.com